Refactor MoH enrollment out of sensor.py; merge moh_extract into moh_native
- Move the ~145-line native-pipeline enroll_moh() implementation out of the
  Sensor class into validitysensor/moh_enrollment.py as a free function
  enroll_moh(sensor, ...). Sensor.enroll_moh is now a thin delegator, so
  existing callers keep working and sensor.py stays device-agnostic.
- Fold moh_extract.py (template serializer, TID derivation, RE scaffold)
  into moh_native.py and delete moh_extract.py, so the MoH pipeline lives
  in a single module.

Lazy imports keep the sensor <-> moh_enrollment dependency acyclic.
